requirements.txt
PyMuPDF
Pillow
numpy
scikit-image
matplotlib
opencv-python
scikit-learn